fix protocol so atm sends request to bank for each command, not stores things in memory

+++ src/bank/bank.h
@@ -33,7 +33,7 @@ typedef struct _Bank
 	unsigned char key[32];
 	HashTable *users;
 	int user_count;
-	int counter;
+	unsigned long counter;
 } Bank;
 
 //struct used to represent a bank user account
